requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asarPy==1.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
